progress
unicode is still horribly fucked tho, it seems as if somehow the pattern
matching is fucking us over. My hypothesis is that the code that does
the thing does not do the thing correctly. It is entirely possible that
that code is in fact bugged, and thus does the thing incorrectly. This
revelation will have drastic impacts on fishing season